Revert "Node == and != operator now compare XIDs"
This reverts commit 27c68f7ec169201d0ff765f270a1f6e88f62edbf.

Had to revert this change to remove ambiguity caused by more than one
Node in the Graph having the same XID. e.g. RV DAG may have same AD
twice.

With this change, if more than one Node had the same XID in the DAG,
the user would have no way of specifying which one they intend to merge
against. So for now the best option is for the user to pre-create all
Node objects and then build Graphs by providing the intended ones as
arguments.