Graph comment explaining Node object usage
Node objects are not compared by value. This lets us create more than
one Node object which represent the same XID appearing more than
once in a Graph. So if the same AD appeared twice and the user wants
to add a path as fallback from the first one, they can specify the
corresponding Node object. If the node objects were compared by value
there would be no way for the user to specify one over the other.