RV service uses Graph for data plane manipulation
The Rendezvous service data-plane used to assume that the intent AD was
the first node in the destination dag in the packet header. So it would
modify that AD after ensuring that the existing XID was an AD. However,
with more complex DAGs we cannot guarantee that the first node will be
the intent AD. So now the Rendezvous service data-plane reads in the
destination DAG as a Graph and calls the library functions to get the
actual intent AD before modifying it in place.

This change also gets rid of a bunch of ugly compiler warning
prevention and a memcpy while making the data-plane more robust.