Providing a way to bypass xcache
If a client does not intend to cache content locally, it can benefit
from improved performance by fetching the content directly instead of
sending its requests via xcache daemon.

The main advantage of this approach is the reduced overhead of
communicating with the xcache daemon over a unix domain socket.

This method will be called if a client requests no caching and is
blocking to wait for the content to arrive. For non-blocking calls and
for caching we still go via xcache daemon.