Write controller_dag to a config file on router
Controllers start with etc/controller_dag containing the DAG of the
controller. When a router joins the controller, the Netjoin daemon also
writes an etc/controller_dag file on the router. So it can be read by
the router when it starts announcing the network for other routers to
join.