RouterCred AD must match announced AD or config AD
RouterCredentials contain the AD that the router wants to join in
response to a network announcement. On a controller, this matches the
configured AD on the system (in address.conf). On a router announcing a
network, there is no configuration file holding the AD. So for now, we
match against the AD that the router was announcing to others. In a
real world setup, this router credential would be forwarded to an
authentication/authorization service for validation.