Restart announcer on rejoining controller
We used to ignore network announcements if we were already announcing a
network. This was to prevent joining other routers in the same network
which was not allowed by the old routing code. However, the new routing
code requires that we allow routers in a network to join other routers.
So disabled the check. Also, when we join a new upstream router, we
should change our announcements to reflect that. This is now done too.