Restart announcer only if controller dag changes
If there is no announcer announcing the network on a router, we start
an announcer. If the network is already being announced, we only
restart the announcer when the controller DAG changes. This allows us
to prevent an endless loop of routers joining their peers. However, the
routers still rejoin their peers if the controller restarts and has a
new DAG.