Use malloc/free to avoid pkt with 0x0 data crash
When a lot of packets are being created with simultaneous XStream
sessions, sometimes the same Packet object was getting queued twice on
the receive queue in XStream for a session. On freeing the first of
those packets, the second one would have 0x0 as its _data pointer.
Later this packet would appear to have a size similar to its memory
address because Packet::size is _tail - _data, with _tail still
pointing to a memory address but _data contains 0x0. Using malloc/free
instead of Packet to store the data avoids this problem. We still have
an occasional deleted Packet being sent to ToDevice but that just gets
dropped and is quite infrequent to cause any problems.