Notify xrouted of foreign networks
Policy module finds out if announced network is local or foreign.
Foreign networks are never joined. Xrouted is notified of them.
New NetjoinXrouted interface between xnetjd and xrouted.
NetjoinSession and NetjoinPolicy use NetjoinXrouted for xrouted
notifications.
Policy module is now started with knowledge of it being a router in
addition to knowing if it was running on a controller.