Add _local_addr on Xupdatedag if AD not known
When non-default interface connects to an upstream router or
controller, we were not setting the _local_addr. With this change, we
look at our _local_addr and if it doesn't include an intent AD, we
accept the newly joined dag as the _local_addr. This will get
overwritten by the _local_addr of the default interface if it connects
to a different network.
NOTE: default routes are not set to match the non-default interface.