Don't set default routes on routers
An additional argument to XupdateDAG, is_router, tells it to not set up
default routes on routers. The default routes will still be set on
hosts. This is required by the new routing code that supports multiple
routers in an AD. Xnetjd was updated accordingly to pass the additional
argument.