Add AD to XARP tables on all interfaces
The AD is added to XARP tables on Controller and Routers. On
Controller, this happens at bootstrap time via assign_addresses.py. On
the routers it is done when a valid handshake two has been received in
response to the router's join request. Note, this may result in
additional idempotent updates when routers join peers or downstream
routers.