Publisher keys and certificate request
This class allows for creation of a Publisher. Each publisher has their
own keypair and can generate a request for a certificate. This request
when presented to a CA would result in the Publisher's certificate that
would allow Clients of Publisher to be able to trust content signed by
the publisher.