Publisher library for NCID operations
New Publisher object represents a publisher by name. Once created, it
is able to access the publisher's keys and use them to create an NCID
for the requested content name.
Also had to make changes to xiasecurity module to provide access to the
SHA1 hash hex representation for forming an NCID.
New Makefile also plugged in to compile and clean the new code.