New NCID and CID header objects
Instead of creating NCID and CID headers on the fly we will have long
lived objects to manage the info with the cache related metadata.
This change introduces the new ContentHeader objects CIDHeader and
NCIDHeader.
In addition, we are moving away from initializing CID information about
in-flight chunks into the metadata. If a chunk is in flight, it could
be an NCID chunk. Only after it arrives completely, we can verify the
contents and store the CID into storage.