ContentHeader replaces cid_header on wire
With this change we are replacing cid_header on wire with
ContentHeaderBuf - a protocol buffer containing a ContentHeader which
may be an NCID or CID header depending on how it was originally created
by the publisher of the content.

Also included is an NCIDTable class that maps between NCIDs and CIDs
and provides atomic add/remove/lookup operations to the mapped
information.