Removing hop_count from xcache code
Based on our understanding, the hop_count from the Xaccept streaming
code in Click stack was exposed to xcache so it could calculate
hop distance of endpoint serving the chunk. This was experimental code
and is no longer needed. Plus it is rather convoluted where the server
must determine the hop count based on the hlim of a packet received
during the handshake. The hlim is set by the client, so deriving
hop_count from it is flawed. The hop_count is then stored for the
latest chunk request only (this was not atomic). The hop_count is then
sent to the client requesting the chunk as part of CID header.
The new design is to have the client work with latency and bandwidth
measurements, which is much simpler and more reliable to work with.