Made click queues huge to avoid overflow.
When a stream client's network is slower than the server side's
connection, the default queue size of 200 caused the server to overflow
the queue. This is mainly an issue due to the fact that the congention
window size can grow to 1m so the server can have a large number of
packets in flight.