removed "context" for store(). sockaddr_x in cmd
__store() used to take a context pointer and it was simply ignored.
This was a problem because I wanted to call the function internally
from xcache without an associated context to download the Publisher
certificate. Simply removed the unused context argument.

xcache_cmd.dag needs a sockaddr_x, but NCID code was storing the dag in
std::string format in it. The compiler never caught it.