Metadata map keys only on CID
Metadata map keys only on CID. If an NCID is provided, the corresponding
CID is used as a key. If the user tries to acquire metadata for an NCID
the corresponding CID is found first.
xcache_meta constructor no longer takes CID as argument. It can be
easily derived from the ContentHeader stored in xcache_meta.