Caching chunks with new content header on routers
Since the content header is no longer fixed size, its size needs to be
fetched first. Followed by the content header, which itself is a
protobuf. Then the size of the chunk needs to be retrieved from the
serialized content header before the rest of the chunk can be
collected. Doing all this while data is flying by in possibly
out-of-order packets is trickier than it looks.
The current implementation assumes that the very first packet must have
the content header, or its initial portion, at least. If that packet
is out-of-order, this code will not collect the chunk. The remaining
packets can be out-of-order and the entire content header doesn't need
to be in the first packet. Although it is unlikely that the content
header will get split into multiple packets.