Memory corruption, class xcache_download
This change includes the following:
1. xcache_download is now a class.
     It used to be a struct that contained a C++ string and with my
changes, additional C++ objects were added to it. Some of these objects
need to be cleaned up in a destructor to simplify memory management.

2. Ongoing downloads are only known to Cache module
     The cache module was adding ongoing downloads to controller
metadata map as soon as one packet was seen for an in-flight chunk.
There is no need for controller to know about a chunk that is not
completely downloaded. We do a cleanup in cache itself. The only catch
is if the chunk is never completed. Need to make sure it will get
cleaned up. One such case is if there is no FIN for a chunk.

3. Memory corruption while downloading content header
      The content header was being copied 4 bytes past the end of
allocated buffer because I failed to account for the content header
size being the first 4 bytes in one "if" statement. Fixed.