XIACidFilter now actually filters CID streams
The XIACidFilter simply forwarded all CID packets to Xcache for
possible in-flight caching on routers. With this change it allows
Xcache to blacklist specific CID streams based on requesting client SID
and the chunk identifier - CID/NCID.
Really, this change provides a communication capability between Xcache
and XIACidFilter in Click to allow for more interesting use cases in
future. Some examples:
* Forwarding non-stream content packets
* Block and unblock specific streams for caching