Return "" on errors getting pubkey.
At a couple of places the return codes were not being checked. In case
there was an error retrieving the public key or certificate for a
publisher, we return "". This empty string should be returned from
calling functions also. Fixed.
The function to read pubkey into a string was not null-terminating the
string. This caused extra characters at the end of pubkey. Fixed.
Also, cleaned up some error output.