If remote fetch fails, server sends empty header
Client calls fetch_content_remote which in turn calls
send_remote_content on the Server. If send_remote_content is unable to
find the content locally, it returns the header size 0 in the first 4
bytes. The Client considers this as a sign that something went wrong at
the server and notifies the XfetchNamedChunk caller that the remote
chunk could not be retrieved.
This condition used to trigger an assert at the server causing Xcache
to crash. However it is easy to trigger this condition if the Client
mistypes the content URI or the requested NCID is not yet available on
the server.