Support caching of pushed chunks
If a chunk is pushed from a publisher to a client, the handshake
direction is reversed. The SYN goes from publisher to client and client
responds with SYNACK to publisher. Compared to SYN going to publisher
and the publisher responding with SYNACK. However, in both transfers,
the data flows in the same direction.

For pushed chunks, we look for SYN going from a CID to an SID.
For pulled chunks, we look for SYNACK going from a CID to an SID.

To make this actually work for pushed chunks, the sender/publisher must
bind to the CID being pushed before pushing it.