Push demo client that accepts a pushed chunk
A simple stream application to accept chunks pushed to it. It does not
do any verification of the received chunk or its header. This is just a
proof-of-concept. An actual implementation would be inside Xcache where
the verification capabilities already exist. The user would call a
(potentially blocking) API and wait for a completely verified chunk to
be delivered.