HACK! Temporary fix for the lone controller issue.
It appears that a controller in an AD with no routers will never update
its routing tables as the code to do so is inside the processLSA function
and we never receive one.
This hack should be done correctly at some point. For now it just
updates the tables just before it sends out the interdomain lsa
message.