Wait only until the chunk is received
The fetch demo client main function was waiting for the received chunk
to be delivered by using a time delay. Now, we use a condition variable
to deliver the chunk to the calling object as soon as it is received.

The problem is again with how Xcache APIs work. The APIs provide for a
callback function to be provided. However, a member function from a
class cannot be such a callback, unless it is defined static. To get
around this, I had to make the pointer of the calling object global
with the callback being a non-member function. Here's how it works:
0. The client waits for chunk by calling FetchDemoClient::get_chunk.
1. We register a non-member function with Xcache API as callback
2. When called back, that function calls FetchDemoClient::got_chunk.
3. FetchDemoClient::got_chunk notifies FetchDemoClient::get_chunk via a
condition variable after safely copying chunk data and making it
available to the caller.