Provide for stopping PushProxy
Once a PushProxy was started there was no way to stop it. That doesn't
quite work very well because I also limited the number of proxies
allowed on a system :). Fixed by making the PushProxy thread
preemptible and the necessary API changes to request XendProxy. The
XnewProxy API had to change to provide a proxy_id/handle to the user
that they can later use to stop the proxy.
Note that one process can have multiple proxies so we can't just use
the context to provide proxy management.