Modernize thread pool code
The thread pool code was using raw pointers instead of smart pointers
that are preferred in the new standards. Also, there was no mechanism
to stop the worker threads. This change introduces an atomic boolean
flag that the threads check from time to time to know if a stop has
been requested. The code is not perfect but works for now.