Moved content headers to api/xcache
Finally able to move content header definitions out of xcache into the
API directory. This enables creation of API that can handle the content
headers without access to xcache. Note that Xcache still maintains an
NCID table that the API does not have access to. This table allows
conversion from NCID to the corresponding underlying CID. However, this
should not prevent creation of an API that can handle and verify
headers for direct downloading of chunks.