Clean up context and notification socket
If the application holding an XcacheHandle is killed with Ctrl-C and
doesn't call XcacheHandleDestroy, the state in Xcache Controller was
not being cleaned up. Xcache Controller keeps a context for each
XcacheHandle that is comprised of a contextID, library socket,
notification socket and a ttl. If the application ends without calling
XcacheHandleDestroy, the Xcache Controller notices that the library
socket was closed. At this point, it now cleans up the context
corresponding to the closed library socket. Earlier this context was
not getting cleanup up causing a leak and file descriptor reuse issues.