Return correct count when no Xsockets in fdsets
If there were no Xsockets in the list of file descriptors being
tracked, we call the native select(). In this case, if there was an
error or timeout, it was reported correctly. However, if some
descriptors were ready, the total count was not returned and the code
would return 0. Fixed.

Also added some safeguards to verify the returned valued. This has some
processing overhead and can be removed if performance is an issue.