Partial fix for hostname mismatches.
The code wasn't truncating the hostname at the first .
An additional fix to deal with hostnames with a - in them has not been
added.
The real solution is to stop using hostnames and just use a fixed name
for the click hostname since it is internal to click and there can only
be one host per machine now.