Adding our own Xfreeifaddrs and Xfreeaddrinfo
We were depending on linux freeifaddrs and freeaddrinfo functions to
clean up memory allocated by our Xgetifaddrs and Xgetaddrinfo functions.
After the multihoming changes, we can no longer depend on the Linux
functions and lose memory if we do. This is because of how we allocate
various chunks of memory that we return to the caller, being different
from how the linux functions do their allocations.