CID/NCID forwarding tables assign local DAG
Instead of ProcessStreamPacket doing special processing to determine if
a SYN packet is destined for a local CID/NCID and updating its DAG to
include Xcache, we do that in forwarding tables for CID/NCID.

New XIACIDRouteTable was created for this purpose. We assume that only
SYN packets for local CID/NCID will make it to the routing table and
only if they are getting delivered to the local stack, we update the
DAG to *->_local_addr->_xcache_sid->cid
where, cid is the requested CID/NCID.

The XIDpair is now client SID and Xcache SID, which we deem unique
enough to identify the chunk being transferred. We no longer use
CID/NCID in the socket pair.

There is additional functionality to notify the forwarding tables of
the current DAG as well as Xcache SID, when they are known during
startup.