If chunk is local, update dest dag in fwd table
The SYN for a chunk request is handled by the forwarding tables for CIDs
and NCIDs. This change allows the forwarding table to update the
destination DAG of SYN requests for chunks that are local. This prevents
the transport from having to check for this situation for every incoming
packet, making it simpler.

This change specifically fixes the implementation of the CID/NCID
forwarding table:
* On routers, the tables were not getting local DAG
* Intent_XID type was in network byte order in XIACIDRouteTable
* The last pointer was pointing to CID instead of Xcache SID
* XIDpair no longer recalculated. Just stored once on connect/accept
* Same XIDpair now used for cleanup
* This allows for XIDpair(XcacheSID, ClientSID) on server side
* CID DAG returned by XputChunk now has no XcacheSID in fallback path
  - Was unnecessary because packet gets routed by HID and CID