temporary hack to add timeout to chunk fetching.
This is not a good general purpose fix and should not make it into the
main line of the codebase. Break out of the loop in the
interruptible_read function if a single Xrecv would take longer than 15
seconds. This should help kill off threads that freeze because an
internal API UDP packet got lost.

Also includes code that lays the groundwork to specify a timeout in the
cache handle but the value is not used in the underlying code yet.