Fixed incorrect usage of sock id in API
The initial implementation for XinterestedInCID allowed for the user to
provide a socket that was not listening for connections and the API was
doing the work of making it appear connected to the application.
However, this didn't play well with Xcache design when it was on the
receiving end. So for now, we'll expect the application to bind and
listen on the socket that is used for XinterestedInCID.