New sockfd wasn't added to active_conns list
The Xcache controller runs a select loop on sockfds listed in
active_conns list. There is a bug with this code that was not cleaning
up a sockfd in the fd_set used by Xselect. However, the corresponding
entry in active_conns was removed. So when a new connection was
created, we recorded it as a duplicate for fd_set. However, the
corresponding entry in active_conns was missing. This caused the
controller to not process the incoming connection on the next Xselect
iteration. Fixed for now by including the duplicate in active_conns
list if it is not already there.