Adding support for multiple SID/HID intents
For mobility, we need to be able to deliver a packet to an alternate
location if the mobile host or service is no longer available. This
change allows for multiple SIDs or multiple HIDs as intents. The change
provides the server side demo applications but the client side
applications will be submitted via a separate change.