sock_type for DGRAM sock set and used for teardown
Xsendto was creating a new sock and setting its source path to reflect
a randomly generated SID. However, the type of socket was not set.

With this change, the sock_type is set for that new socket. The same is
later used by TeardownSocket to remove the route to the randomly
generated SID.