XID calculated for deleting route was incorrect
With multiple-SID DAGs like this one:
    *-> AD -> HID ---------> SID1
               '--> SID2 -'

The two sockets will bind to:
    *-> AD -> HID -> SID1
and *-> AD -> HID -> SID2

But the accepted socket for SID2 will use this address:
    *-> AD -> HID -> SID2 -> SID1

and the accepted socket for SID1 will use the original complicated DAG.

So, when we go to TeardownSocket, it needs to delete routes only for
SID1 and SID2 only when their listening sockets close. But at the same
time remove book-keeping entries for the pairs (SID2, SIDclient) and
(SID1, SIDclient) even though the intent for SID2 DAG is SID1.
Also, forgot to account for how all this affects DGRAM sockets!

Next, see if this code still makes sense for the multiple HID use-case.