Move GNS registration out to a header file
This changeset moves most of the GNS interaction details out of the
applications and into a new header: Xgns.h

The GNSServer instance connects to a local GNSHTTPProxy instance which
is currently hard-coded to run at port 5678 on localhost.
It then allows for creation and removal of entries in the GNSServer.
Note: GNSServer object functions are not thread-safe at this time.