Modify initial scenario settings.
Limit total bandwidth available from 4 of the 6 CDNs in the default
config. This should force the broker to switch CDNs assigned to a
particular client more often.

Once under load, the broker is less willing to change CDNs for a client.
It does change more often betwen load tests however. I think this is due
to the broker algorithms which didn't take into account how we are using
it now.