IPaddr:port for nexthop support in bin/xroute
If the next hop is an IP address and port number, send that info to
bin/xroute which can then display it to the user. Also, allow the user
to create a new entry with IP addr:port as the next hop. The code to
delete the entry has not been updated in this change.