Infrastructure to run XIA as an overlay
This introduces two utilities:
1. xiaconfigurator.py - takes a conf file and configures all routers
2. xiaconfighelper.py - helper running on each router to aid in config

The communication between the above is via protocol buffers.
There is no Makefile to compile the proto spec file yet.

As of this writing, info is gathered from each router and a router.click
file is deployed on each router. Next, we need to launch xianet on each
router and then gather AD, HID info. That will then be used to set up
the routes based on topology configured by user.