Gather route info and build route commands
We now read the static routing configuration from the config file and
build corresponding forwarding table setup commands to be deployed on
the individual routers. The routes are not configured on the routers by
this commit. They are just displayed. The configuration will happen
after we verify that the routes are being generated as expected.