XIAOverlayFilter needs to uniqueify outgoing pkt.
A packet leaving XIAOverlayFilter in the XIALineCard needs to be
uniqueify() because it is fed into a Tee element followed by a Queue.
The cloning in Tee() followed by one of the clones being processed
results in a memory corruption in the data shared with the packet that
is still heading to the Queue. This resulted in the XIA header in the
packet data having 4 bytes trampled.