Stopping XIA is a best effort command
In case the router is not configured yet, the stop command will fail and
that is OK because this is followed by an attempt to start XIA where we
do check to ensure that it succeeded. If the stop fails the following
start XIA command will also fail because click would be already running.