Allow configuration to define where Xcache runs
With this change, each router can have an Xcache entry in the
configuration file that tells the configurator if an Xcache instance
should be started on the router. For now, the Xcache process does not
take any arguments. If they are needed in future, we could add them to
the config file too and change the parsing accordingly. Alternatively,
the Xcache on each router uses its own xcache.local.conf file and any
static configuration can be placed there if changes to this configurator
are undesirable.

Also, note that we made a hard assumption that the picoquic directory
will exist parallel to the xia-core directory:
+--> xia-core
`--> picoquic

The exception handler that produced a generic message was disabled. This
will allow us to debug potential issues faster because the backtrace
would be available for debugging without having to run the configurator
again with the exception handler commented.